Don't defer resizes if the current thread matches
Attempt to track the "home thread" for a surface, and only do resizes
(which involve resetting the EGLContext) on that thread. While this
isn't a full solution -- it's still possible that the old
surface/context will be bound on other threads if the application uses
several -- it is an improvement in many cases. In particular,
single-threaded applications (or, at least, applications which interact
with wayland and egl from the same thread) won't lose the contents of
their surface just before presenting.

This still has a regression in Life is Strange on KDE: the fullscreen
transision doesn't quite go through as planned, resulting in the window
not fully covering the screen. It doesn't crash, though, and force
toggling "Fullscreen" from KDE's menu can fix it.